+++ CHANGES.md
@@ -5,6 +5,7 @@
 ### Github Issues
 - #122: Set horizontal layout direction based on the language
 - #698: Add support for MUC private messages
+- #1038: Support setting node config manually
 - #1057: Removed the `mobile` view mode. Instead of setting `view_mode` to `mobile`, set it to `fullscreen`.
 - #1174: Show MUC avatars in the rooms list
 - #1195: Add actions to quote and copy messages